Some code review cleanup
