+++ include/onnxruntime/core/framework/data_types_internal.h
@@ -13,7 +13,7 @@
 #include "boost/mp11.hpp"
 
 #include "core/common/common.h"
-#ifndef PROVIDER_BRIDGE_PROVIDER
+#ifndef SHARED_PROVIDER
 #include "core/common/type_list.h"
 #include "core/framework/data_types.h"
 #include "core/graph/onnx_protobuf.h"
